feat limit schedule reorder
limit it to rers, and make it a de-activated option by default.
other schedules do not have the same format, so let's not take too much 
risk.
I might have issues for alphabetic message which are not times, if it's 
not the next train...
add a counter for reordering in debug mode to evaluate the need